requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==1.5.3
matplotlib==3.7.2
python-dotenv==1.0.0